+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 9.0 (2020-03-31)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/8.0...9.0
+
+- 129 Jobs postcode search
+- 169 Adjust position of form field help text
+- Forms integration with Aptean Respond API
+
 ## 8.0 (2020-03-30)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/7.0...8.0